finish naive tree

+++ packages/naive-tree/db/bad/schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE Comments (
   comment_id SERIAL,
-  parent_id BIGINT,
+  parent_id INTEGER,
   comment TEXT NOT NULL,
   PRIMARY KEY (comment_id),
   FOREIGN KEY (parent_id) REFERENCES Comments(comment_id)
